Safari-609.1.11.1/LayoutTests/css3/shapes/shape-outside/values/shape-outside-circle-008-expected.txt


FAIL circle(at foo) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 50) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 10px 20px 30px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 20% 30% 40%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 20px, 30px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 20%, 30%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 20px, 30px, 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 20%, 30%, 40%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at closest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at farthest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 20px 30px closest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 20px 30px farthest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 20px 30px foo) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at closest-side 20px 30px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at farthest-side 20px 30px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 20% 30% closest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at 20% 30% farthest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at closest-side 20% 30%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at farthest-side 20% 30%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at closest-side farthest-side 20px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at closest-side farthest-side 20%) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at farthest-side closest-side closest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at farthest-side, closest-side, closest-side) - invalid assert_equals: expected (object) null but got (string) ""
FAIL circle(at left 50x top 50px bottom) - invalid assert_equals: expected (object) null but got (string) ""
